Switch to GNU AFFERO
Changed license to GNU AFFERO 3.0. 
This has the same limitations of the MIT license, the same permissions (plus the patent use, that is now included in the permissions) and adds the following conditions to the software:
- State changes
- Disclose source
- Network use is distribution
- Same license

This license is effective from now on, in every version released after 3.3.96. Older versions will still be available under the MIT license.

Signed-off-by: Filippo Romani <filippo@romani.cc>